Fixed compile errors

+++ src/entity.cpp
@@ -5,12 +5,12 @@ Entity::Entity() : position_x(0), position_y(0), boundary_x(0), boundary_y(0), o
 	
 }
 // Basic Constructor (excludes offsets)
-Entity::Entity(float px, float py, float bx, float by) : position_x(px), position_y(py), boundary_x(bx), boundary_y(by), offset_x(0), offset_y(0)
+Entity::Entity(double px, double py, double bx, double by) : position_x(px), position_y(py), boundary_x(bx), boundary_y(by), offset_x(0), offset_y(0)
 {
 	
 }
 // Full Constructor (includes offsets)
-Entity::Entity(float px, float py, float bx, float by, float ox, float oy) : position_x(px), position_y(py), boundary_x(bx), boundary_y(by), offset_x(ox), offset_y(oy)
+Entity::Entity(double px, double py, double bx, double by, double ox, double oy) : position_x(px), position_y(py), boundary_x(bx), boundary_y(by), offset_x(ox), offset_y(oy)
 {
 	
 }
@@ -27,27 +27,27 @@ void Entity::collide(Entity* ent)
 {
 
 }
-float Entity::get_position_x()
+double Entity::get_position_x()
 {
 	return position_x;
 }
-float Entity::get_position_y()
+double Entity::get_position_y()
 {
 	return position_y;
 }
-float Entity::get_boundary_x()
+double Entity::get_boundary_x()
 {
 	return boundary_x;
 }
-float Entity::get_boundary_y()
+double Entity::get_boundary_y()
 {
 	return boundary_y;
 }
-float Entity::get_offset_x()
+double Entity::get_offset_x()
 {
 	return offset_x;
 }
-float Entity::get_offset_y()
+double Entity::get_offset_y()
 {
 	return offset_y;
 }
